requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
Django==2.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
kombu==4.2.1
pytz==2018.5
redis==2.10.6
vine==1.1.4
